Moved tests to describe_task_queue to be consistent with the current style
